updated package.json to run launch file at firstly
