+++ package.json
@@ -13,9 +13,9 @@
     "watch-json": "npm run copy-configs -w",
     "copy-configs": "cpx \"src/configurations/*.json\" build/configurations",
     "launch": "node build/app.js",
-    "build-launch": "npm run build & npm run launch",
+    "build-launch": "npm run build | npm run launch",
     "tslint": "tslint -c tslint.json -p tsconfig.json",
-    "test": "mocha -r ts-node/register test/**/**.test.ts"
+    "test": "mocha -r ts-node/register test/launch.ts"
   },
   "keywords": [],
   "author": "",
