+++ src/types/server.ts
@@ -1,9 +1,7 @@
 import { Server as HapiServer } from 'hapi';
-import { Sequelize } from 'sequelize';
-import { Response } from './response';
+import { IBoomMethods } from './response';
 
 export interface Server extends HapiServer {
-    db?: () => Sequelize;
-    boom(): Response;
+    boom(): IBoomMethods;
 }
