+++ test/launch.ts
@@ -1,2 +1,2 @@
-// require('./file/api.test');
+require('./file/api.test');
 require('./user/api.test');
